Move struct related methods to its own class
Also rename thrift_util to thrift_message and add some unit test for the
thrift_struct class
